pyproject.toml
[tool.poetry]
name = "animals-system-hw23"
version = "0.1.0"
description = ""
authors = ["qweArteom <kurylenko2artem@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
